backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
openai==1.54.0
httpx>=0.24,<0.29
numpy==1.26.4
pydantic>=2.9.2
google-generativeai==0.8.3
supabase>=2.25.0
websockets>=13.0



